+++ node/src/main/kotlin/net/corda/node/services/schema/NodeSchemaService.kt
@@ -66,11 +66,12 @@ class NodeSchemaService(extraSchemas: Set<MappedSchema> = emptySet(), includeNot
     object NodeNotary
 
     object NodeNotaryV1 : MappedSchema(schemaFamily = NodeNotary.javaClass, version = 1,
-            mappedTypes = listOf(PersistentUniquenessProvider.PersistentUniqueness::class.java,
-                    PersistentUniquenessProvider.PersistentNotaryCommit::class.java,
-                    RaftUniquenessProvider.RaftState::class.java,
-                    BFTNonValidatingNotaryService.PersistedCommittedState::class.java
-            )){
+            mappedTypes = listOf(PersistentUniquenessProvider.BaseComittedState::class.java,
+                    PersistentUniquenessProvider.Request::class.java,
+                    PersistentUniquenessProvider.CommittedState::class.java,
+                    RaftUniquenessProvider.CommittedState::class.java,
+                    BFTNonValidatingNotaryService.CommittedState::class.java
+            )) {
         override val migrationResource = "node-notary.changelog-master"
     }