+++ node/src/main/kotlin/net/corda/node/services/transactions/SimpleNotaryService.kt
@@ -18,7 +18,7 @@ import java.security.PublicKey
 
 /** A simple Notary service that does not perform transaction validation */
 class SimpleNotaryService(override val services: ServiceHubInternal, override val notaryIdentityKey: PublicKey) : TrustedAuthorityNotaryService() {
-    override val uniquenessProvider = PersistentUniquenessProvider()
+    override val uniquenessProvider = PersistentUniquenessProvider(services.clock)
 
     override fun createServiceFlow(otherPartySession: FlowSession): NotaryFlow.Service = NonValidatingNotaryFlow(otherPartySession, this)